no separate serverside and clientside components
